+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 - LLVM for prebuilt packages bumped to v15.0.7. (#4311)
 - Linker-level dead code elimination is enabled by default for Apple, wasm and *all* ELF targets too now. (#4320)
 - Vector comparisons (==, !=, <, <=, >, >=) now yield a vector mask. Identity comparisons (`is`, `!is`) still yield a scalar `bool`. (3a59ee81)
+- New `timetrace2txt` tool for easier inspection of `--ftime-trace` output. (#4335)
 
 #### Platform support
 - Supports LLVM 9.0 - 15.0.